samples: sensor: add sensor sample for TCN75A alert
Add sample for TCN75A, which demonstrates the alert feature. The
application will read the current temperature, and then set the target
lower and upper limit to 0.5 C and 1.5 C above the ambient temperature.
The user can warm the sensor up and allow it to cool down to observe the
alert functionality.

Signed-off-by: Daniel DeGrasse <daniel@degrasse.com>