wip: enabling driver for IS31FL LED matrix driver (with matrix logic)
Added LED matrix driver, which implements the majority of the matrix
logic in the driver code. This allows the user to use LED indices based
off their chosen array size, but reduces flexibility for alternative
use cases (for example, if the user wanted to split the 12x16 array into
two 12x8 arrays, and stack them to emulate a 24x7 array.

This driver will likely not be upstreamed in this form, because the
logic of calculating LED indicies based on custom matrix layouts should
likely be left to the application

Current driver states:
- led on/off works
- led brightness setting works
- led_write_channels works, sets PWM value for all LEDs
- auto breathe mode is not implemented.

Signed-off-by: Daniel DeGrasse <daniel@degrasse.com>