wip: new LED algorithm implementation
Implemented new LED algorithm, which calculates the LED delay
significatly faster (at least on an M4 core with an FPU)

This algorithm uses log2 functions, which requires newlib nano or
picolibc. Still trying to figure out how to indicate this in the driver
Kconfig settings.

Also, this algorithm needs to be tested more throughly. Basic values
have been tested, but running this on a desktop across a range of ramp
up, ramp down, and on/off values to verify it returns the same values
would be useful

Signed-off-by: Daniel DeGrasse <daniel@degrasse.com>